+++ internal/resolvers/twitch/user_loader.go
@@ -23,6 +23,17 @@ type twitchUserTooltipData struct {
 	URL         string
 }
 
+type twitchUserLiveTooltipData struct {
+	Name        string
+	CreatedAt   string
+	Description string
+	URL         string
+	Title       string
+	Game        string
+	Viewers     string
+	Uptime      string
+}
+
 type UserLoader struct {
 	helixAPI TwitchAPIClient
 }
@@ -50,15 +61,25 @@ func (l *UserLoader) Load(ctx context.Context, login string, r *http.Request) (*
 
 	var user = response.Data.Users[0]
 
-	var name string
-	if strings.ToLower(user.DisplayName) == login {
-		name = user.DisplayName
-	} else {
-		name = fmt.Sprintf("%s (%s)", user.DisplayName, user.Login)
+	streamResponse, err := l.helixAPI.GetStreams(&helix.StreamsParams{UserLogins: []string{login}})
+	if err != nil || len(streamResponse.Data.Streams) == 0 {
+		return userResponse(login, user)
 	}
 
+	return userLiveResponse(login, user, streamResponse.Data.Streams[0])
+}
+
+func buildName(login string, user helix.User) string {
+	if strings.ToLower(user.DisplayName) == login {
+		return user.DisplayName
+	} else {
+		return fmt.Sprintf("%s (%s)", user.DisplayName, user.Login)
+	}
+}
+
+func userResponse(login string, user helix.User) (*resolver.Response, time.Duration, error) {
 	data := twitchUserTooltipData{
-		Name:        name,
+		Name:        buildName(login, user),
 		CreatedAt:   humanize.CreationDate(user.CreatedAt.Time),
 		Description: user.Description,
 		URL:         fmt.Sprintf("https://twitch.tv/%s", user.Login),
@@ -73,5 +94,32 @@ func (l *UserLoader) Load(ctx context.Context, login string, r *http.Request) (*
 		Status:    200,
 		Tooltip:   url.PathEscape(tooltip.String()),
 		Thumbnail: user.ProfileImageURL,
+	}, cache.NoSpecialDur, nil
+}
+
+func userLiveResponse(login string, user helix.User, stream helix.Stream) (*resolver.Response, time.Duration, error) {
+	data := twitchUserLiveTooltipData{
+		Name:        buildName(login, user),
+		CreatedAt:   humanize.CreationDate(user.CreatedAt.Time),
+		Description: user.Description,
+		URL:         fmt.Sprintf("https://twitch.tv/%s", user.Login),
+		Title:       stream.Title,
+		Game:        stream.GameName,
+		Viewers:     humanize.Number(uint64(stream.ViewerCount)),
+		Uptime:      humanize.Duration(time.Since(stream.StartedAt)),
+	}
+
+	var tooltip bytes.Buffer
+	if err := twitchUserLiveTooltip.Execute(&tooltip, data); err != nil {
+		return resolver.Errorf("Twitch user template error: %s", err)
+	}
+
+	thumbnail := strings.ReplaceAll(stream.ThumbnailURL, "{width}", "1280")
+	thumbnail = strings.ReplaceAll(thumbnail, "{height}", "720")
+
+	return &resolver.Response{
+		Status:    200,
+		Tooltip:   url.PathEscape(tooltip.String()),
+		Thumbnail: thumbnail,
 	}, cache.NoSpecialDur, nil
 }